Moving changes from the 'dao' to the 'appendonlydao' package - ContractStateEventsReader
