+++ ledger/participant-integration-api/src/main/scala/platform/store/dao/LedgerDao.scala
@@ -26,6 +26,7 @@ import com.daml.lf.data.Ref.{PackageId, Party}
 import com.daml.lf.transaction.BlindingInfo
 import com.daml.logging.LoggingContext
 import com.daml.platform.indexer.OffsetStep
+import com.daml.platform.store.dao.events.{ContractStateEvent, FilterRelation, TransactionsWriter}
 import com.daml.platform.store.dao.events.TransactionsWriter.PreparedInsert
 import com.daml.platform.store.dao.events.{FilterRelation, TransactionsWriter}
 import com.daml.platform.store.entries.{
@@ -76,7 +77,7 @@ private[platform] trait LedgerDaoTransactionsReader {
       endInclusive: (Offset, Long),
   )(implicit
       loggingContext: LoggingContext
-  ): Source[((Offset, Long), events.ContractStateEventsReader.ContractStateEvent), NotUsed]
+  ): Source[((Offset, Long), ContractStateEvent), NotUsed]
 }
 
 private[platform] trait LedgerDaoCommandCompletionsReader {
