+++ ledger/participant-integration-api/src/test/suite/scala/platform/store/dao/JdbcLedgerDaoH2DatabaseSpec.scala
@@ -23,4 +23,3 @@ final class JdbcLedgerDaoH2DatabaseSpec
     with JdbcLedgerDaoTransactionTreesSpec
     with JdbcLedgerDaoTransactionsWriterSpec
     with JdbcAtomicTransactionInsertion
-    with JdbcLedgerDaoContractEventsStreamSpec
