+++ ledger/participant-integration-api/src/test/suite/scala/platform/store/dao/JdbcLedgerDaoPostgresqlSpec.scala
@@ -23,4 +23,3 @@ final class JdbcLedgerDaoPostgresqlSpec
     with JdbcLedgerDaoTransactionTreesSpec
     with JdbcLedgerDaoTransactionsWriterSpec
     with JdbcAtomicTransactionInsertion
-    with JdbcLedgerDaoContractEventsStreamSpec
